requirements.txt
websocket_client==0.57.0
pycryptodome==3.9.4
qrcode[pil]==6.1